backend/requirements.txt
Django>=5.1.0,<5.2.0Django==5.1.3

djangorestframework>=3.15.0djangorestframework==3.15.2

djangorestframework-simplejwt>=5.3.0djangorestframework-simplejwt==5.3.1

django-cors-headers>=4.5.0django-cors-headers==4.5.0

drf-yasg>=1.21.0psycopg2-binary==2.9.10

psycopg2-binary>=2.9.0python-dotenv==1.0.1

python-dotenv>=1.0.0
Pillow>=10.0.0
